requirements-dev.txt
black==25.1.0
cachetools==5.5.2
certifi==2025.4.26
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.1.8
coverage==7.8.0
db-dtypes==1.4.2
distlib==0.3.9
filelock==3.18.0
google-api-core==2.24.2
google-auth==2.40.1
google-auth-oauthlib==1.2.2
google-cloud-bigquery==3.31.0
google-cloud-bigquery-storage==2.31.0
google-cloud-core==2.4.3
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
identify==2.6.10
idna==3.10
iniconfig==2.1.0
isort==6.0.1
mypy_extensions==1.1.0
nodeenv==1.9.1
numpy==2.2.5
oauthlib==3.2.2
packaging==25.0
pandas==2.2.3
pandas-gbq==0.28.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.5.0
pre_commit==4.2.0
proto-plus==1.26.1
protobuf==6.31.0rc2
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydata-google-auth==1.9.1
pytest==8.3.5
pytest-cov==6.1.1
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
scipy==1.15.2
setuptools==80.3.1
six==1.17.0
tqdm==4.67.1
tzdata==2025.2
urllib3==2.4.0
virtualenv==20.31.1
